allow the use of differents tsConfig file name
When using ts-config-path the name must be tsconfig.json, if you use a
different one it will fallback due ts-loader to the default
tsconfig.json.

Ts-loader doesn't have a direct way to setup a full path for the
TsConfigFile, but has a hack consiting to get special environment
variable TS_NODE_PROJECT to get the filename.

when using simple and lite webpack config file, we simply want to set
the stuff directly through the configuration options.

Signed-off-by: sebastien colleur <sebastien.colleur@gmail.com>